requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
beautifulsoup4==4.6.0
boto3==1.4.1
botocore==1.4.93          # via boto3, s3transfer
cachetools==2.1.0         # via zenpy
certifi==2018.1.18        # via requests
chardet==3.0.4            # via requests
directory-client-core==4.3.0  # via directory-cms-client, directory-forms-api-client
directory-cms-client==8.2.0
directory-components==3.5.0
directory-constants==11.1.0
directory-forms-api-client==3.1.0
directory-healthcheck==0.5.1
directory-validators==5.0.0
django-admin-ip-restrictor==1.0.0  # via directory-components
django-crispy-forms==1.7.1
django-environ==0.4.5
django-formtools==2.0
django-health-check==3.0.0  # via directory-healthcheck
django-ipware==2.1.0      # via django-admin-ip-restrictor
django-recaptcha==1.2.0
django-redis==4.8.0
django-storages==1.5.1
django==1.11.13
djangorestframework==3.7.7  # via sigauth
docutils==0.14            # via botocore
export-elements==0.23.0
gunicorn==19.5.0
idna==2.6                 # via requests
jmespath==0.9.3           # via boto3, botocore
jsonschema==2.6.0
mohawk==0.3.4             # via directory-components, sigauth
monotonic==1.4            # via directory-client-core
olefile==0.44             # via directory-validators, pillow
phonenumberslite==7.7.4   # via directory-validators
pillow==4.2.1             # via directory-validators
python-dateutil==2.6.1    # via botocore, zenpy
pytz==2017.2              # via directory-healthcheck, directory-validators, django
raven==6.9.0
redis==2.10.6             # via django-redis
requests==2.18.4
s3transfer==0.1.13        # via boto3
sigauth==4.0.1            # via directory-client-core
six==1.11.0               # via mohawk, python-dateutil, w3lib
urllib3==1.22             # via requests
w3lib==1.19.0             # via directory-client-core
waitress==0.8.9
whitenoise==3.3.1
zenpy==1.1.10
